+++ pkg/ruler/api_test.go
@@ -3,6 +3,7 @@ package ruler
 import (
 	"context"
 	"encoding/json"
+	io "io"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -25,8 +26,7 @@ func TestRuler_rules(t *testing.T) {
 	defer rcleanup()
 	defer services.StopAndAwaitTerminated(context.Background(), r) //nolint:errcheck
 
-	req := httptest.NewRequest("GET", "https://localhost:8080/api/prom/api/v1/rules", nil)
-	req.Header.Add(user.OrgIDHeaderName, "user1")
+	req := requestFor(t, "GET", "https://localhost:8080/api/prom/api/v1/rules", nil, "user1")
 	w := httptest.NewRecorder()
 	r.PrometheusRules(w, req)
 
@@ -81,8 +81,7 @@ func TestRuler_rules_special_characters(t *testing.T) {
 	defer rcleanup()
 	defer services.StopAndAwaitTerminated(context.Background(), r) //nolint:errcheck
 
-	req := httptest.NewRequest("GET", "https://localhost:8080/api/prom/api/v1/rules", nil)
-	req.Header.Add(user.OrgIDHeaderName, "user1")
+	req := requestFor(t, http.MethodGet, "https://localhost:8080/api/prom/api/v1/rules", nil, "user1")
 	w := httptest.NewRecorder()
 	r.PrometheusRules(w, req)
 
@@ -137,8 +136,7 @@ func TestRuler_alerts(t *testing.T) {
 	defer rcleanup()
 	defer r.StopAsync()
 
-	req := httptest.NewRequest("GET", "https://localhost:8080/api/prom/api/v1/alerts", nil)
-	req.Header.Add(user.OrgIDHeaderName, "user1")
+	req := requestFor(t, http.MethodGet, "https://localhost:8080/api/prom/api/v1/alerts", nil, "user1")
 	w := httptest.NewRecorder()
 	r.PrometheusAlerts(w, req)
 
@@ -192,20 +190,56 @@ rules:
 	router.Path("/api/v1/rules/{namespace}/{groupName}").Methods("GET").HandlerFunc(r.GetRuleGroup)
 
 	// POST
-	req := httptest.NewRequest("POST", "https://localhost:8080/api/v1/rules/namespace", strings.NewReader(rules))
-	req.Header.Add(user.OrgIDHeaderName, "user1")
-	ctx := user.InjectOrgID(req.Context(), "user1")
+	req := requestFor(t, http.MethodPost, "https://localhost:8080/api/v1/rules/namespace", strings.NewReader(rules), "user1")
 	w := httptest.NewRecorder()
 
-	router.ServeHTTP(w, req.WithContext(ctx))
+	router.ServeHTTP(w, req)
 	require.Equal(t, 202, w.Code)
 
 	// GET
-	req = httptest.NewRequest("GET", "https://localhost:8080/api/v1/rules/namespace/test", nil)
-	req.Header.Add(user.OrgIDHeaderName, "user1")
-	ctx = user.InjectOrgID(req.Context(), "user1")
+	req = requestFor(t, http.MethodGet, "https://localhost:8080/api/v1/rules/namespace/test", nil, "user1")
 	w = httptest.NewRecorder()
 
-	router.ServeHTTP(w, req.WithContext(ctx))
+	router.ServeHTTP(w, req)
 	require.Equal(t, 200, w.Code)
+}
+
+func TestRuler_DeleteNamespace(t *testing.T) {
+	cfg, cleanup := defaultRulerConfig(newMockRuleStore(mockRulesNamespaces))
+	defer cleanup()
+
+	r, rcleanup := newTestRuler(t, cfg)
+	defer rcleanup()
+	defer services.StopAndAwaitTerminated(context.Background(), r) //nolint:errcheck
+
+	router := mux.NewRouter()
+	router.Path("/api/v1/rules/{namespace}").Methods(http.MethodDelete).HandlerFunc(r.DeleteNamespace)
+	router.Path("/api/v1/rules/{namespace}/{groupName}").Methods(http.MethodGet).HandlerFunc(r.GetRuleGroup)
+
+	// Verify namespace1 rules are there.
+
+	req := requestFor(t, http.MethodGet, "https://localhost:8080/api/v1/rules/namespace1/group1", nil, "user1")
+	w := httptest.NewRecorder()
+
+	router.ServeHTTP(w, req)
+	require.Equal(t, http.StatusOK, w.Code)
+	require.Equal(t, "name: group1\ninterval: 1m\nrules:\n    - record: UP_RULE\n      alert: \"\"\n      expr: up\n    - record: \"\"\n      alert: UP_ALERT\n      expr: up < 1\n", w.Body.String())
+
+	// Delete namespace1
+	req = requestFor(t, http.MethodDelete, "https://localhost:8080/api/v1/rules/namespace1", nil, "user1")
+	w = httptest.NewRecorder()
+
+	router.ServeHTTP(w, req)
+	require.Equal(t, http.StatusAccepted, w.Code)
+	require.Equal(t, "{\"status\":\"success\",\"data\":null,\"errorType\":\"\",\"error\":\"\"}", w.Body.String())
+}
+
+func requestFor(t *testing.T, method string, url string, body io.Reader, userID string) *http.Request {
+	t.Helper()
+
+	req := httptest.NewRequest(method, url, body)
+	req.Header.Add(user.OrgIDHeaderName, userID)
+	ctx := user.InjectOrgID(req.Context(), userID)
+
+	return req.WithContext(ctx)
 }